+++ shared-rust/shared-frontend/src/components/mod.rs
@@ -5,4 +5,5 @@
 
 pub mod footer;
 pub mod header;
+pub mod language_switcher;
 pub mod notifier;